Reverted Vectrex core to 8 bit audio

+++ VecXGL/VecXGL/VectrexGameCore.h
@@ -36,7 +36,7 @@ OE_EXPORTED_CLASS
     uint8_t yAxis[2][2];
 }
 
-- (void)updateSound:(uint16_t *)buff len:(int)len;
+- (void)updateSound:(uint8_t *)buff len:(int)len;
 
 @end
